requirements.txt
feedparser
google-generativeai
python-dotenv
requests
python-telegram-bot
sqlite3